iraira/pyproject.toml
[tool.poetry]
name = "iraira"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.9, <=3.10"
readchar = "^3.0.5"
numpy = "^1.22.3"

pyaudio = [
    # apt install 
    {version = "^0.2.11", markers = "platform_system == 'Linux'" },

    # Windows binary: Python 3.10 unofficial binary https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio
    {path = "./libs/PyAudio-0.2.11-cp39-cp39-win_amd64.whl", markers = "platform_system == 'Windows'", python = "~3.9"},
    {path = "./libs/PyAudio-0.2.11-cp310-cp310-win_amd64.whl", markers = "platform_system == 'Windows'", python = "~3.10"},
]

"RPi.GPIO" = [
    {version = "^0.7.1", markers = "platform_system == 'Linux'"},
]
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.950"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
